% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotype_boxplots.R
\name{generate_boxplots}
\alias{generate_boxplots}
\title{Genotype boxplots for STOAT GWAS Results}
\usage{
generate_boxplots(phenotype_file, dir_path, output_path)
}
\description{
Generates boxplots of phenotype by inferred genotype.
}
