% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotype_boxplots.R
\name{genotype_boxplots}
\alias{genotype_boxplots}
\title{Genotype boxplots for STOAT GWAS Results}
\usage{
genotype_boxplots(phenotype_file, dir_path, output_path = "output_boxplots")
}
\arguments{
\item{phenotype_file}{Path to the phenotype file use for the GWAS analysis.}

\item{dir_path}{Directory containing N-table files with genotype}

\item{output}{Path to save the output plot image.}
}
\value{
Saves a genotype boxplots to the specified file.
}
\description{
Generates boxplots of phenotype by inferred genotype.
}
