% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manhattan_plot.R
\name{manhattan_plot}
\alias{manhattan_plot}
\alias{plot_manhattan_quantitative}
\title{Manhattan Plots for STOAT GWAS Results}
\usage{
plot_manhattan_quantitative(file_path, output_manhattan = "manhattan_plot.png")
}
\description{
Generate Manhattan plots for quantitative or binary for STOAT GWAS results.
}
