% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths_plot.R
\name{plot_path_length_dotplot}
\alias{plot_path_length_dotplot}
\title{Dot Plot of Path Length Frequencies}
\usage{
plot_path_length_dotplot(input, output = "paths_length_distribution_plot.png")
}
\arguments{
\item{input}{Path to the input TSV file.}

\item{output}{Path to save the output plot image.}
}
\value{
Saves a dot plot to the specified file.
}
\description{
Create a dot plot from a TSV file containing a `path_length` column.
}
