% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qq_plot.R
\name{qq_plot}
\alias{qq_plot}
\title{Q-Q Plot for GWAS Results}
\usage{
qq_plot(input, column_names = "", output = "qq_plot.png")
}
\arguments{
\item{input}{Path to the input TSV file (must contain a column named 'P').}

\item{column_names}{Column name to use for p-values (default: ""). If empty, will use "P" or "P_CHI2" if available.}

\item{output}{Filename for the output PNG plot (default: "qq_plot.png").}
}
\value{
Saves a Q-Q plot image.
}
\description{
Generate QQ plot from STOAT GWAS results using P or P_CHI2 column.
}
