% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotype_boxplots.R
\name{genotype_boxplots}
\alias{genotype_boxplots}
\title{Genotype boxplots for STOAT GWAS Results}
\usage{
genotype_boxplots(genotype_file, phenotype_file, output_path)
}
\arguments{
\item{input}{Path to the input GWAS TSV file.}

\item{output}{Path to save the output plot image.}
}
\value{
Saves a genotype boxplots to the specified file.
}
\description{
Generates boxplots of phenotype by inferred genotype.
}
