% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manhattan_plot.R
\name{manhattan_plot}
\alias{manhattan_plot}
\title{Manhattan Plot for GWAS Results}
\usage{
manhattan_plot(input, output_manhattan = "manhattan_plot.png")
}
\arguments{
\item{input}{Path to the input GWAS TSV file.}

\item{output}{Path to save the output plot image.}
}
\value{
Saves a manhattan_plot to the specified file.
}
\description{
Generate Manhattan plots from STOAT GWAS results (keeps CHR names like 'chr1', 'chrX', etc.)
}
